backend/requirements.txt
odmantic
fastapi==0.109.0
pydantic
uvicorn[standard]==0.27.0.post1
python-dotenv
pandas
PyPDF2
pdfminer.six
httpx
python-multipart
requests
tika
rakun2
bokeh
thefuzz
rapidfuzz
passlib
python-jose
scikit-learn
pytest
mongomock
pytest-asyncio
httpx
bcrypt